The first attempt at finding possible rainbow area, using the EUMETSAT satellite
This is not the satellite we’ll use in the end,
but it shows the principle: taking an image of the clouds, distorting it
using a barrel distortion with the sun at its center, then overlay this
on the original image to find the edges that are hit by the sun.

Areas with heightened rainbow chance!

Also added a logging framework and moved the geo functions into their own file.