Find the latest forecast on precitable water on the Global Forecast System, and draw it.
Thanks on the pointers on Cameron Beccario’s https://github.com/cambecc/earth/ we knew
how to get to the GRIB files and convert them to JSON (with grib2json).

The ‘fetch’ script will try to get the latest forecast, going back six hours at a time right until it finds
a file it can download.

A simple PIL function plots the data.