The logger has two possible behaviours, based on the DEBUG setting.
If DEBUG is true, the logger prints to stdout.
See: http://stackoverflow.com/questions/2302315/how-can-info-and-debug-logging-message-be-sent-to-stdout-and-higher-level-messag#answer-9323805

If DEBUG is false, the logger logs to a MongoDB database.
In this way, it is easy to keep all the logs in one place,
and show them in a web-based monitoring application,
available through app.py at the uri /hq/

added a file for local settings