Check which cities are in one of the rainbow_areas
Checking from the rainbow_areas predicted in water.py

position_to_point allows to translate a geographic coordinate into a pixel value,
for the image we produced from the gfs data:
there are black pixels where we predict rainbows.

We get the cities from a list we constructed on basis of Wikipedia / Russian census data.
We had to remove two entries because they do not have location data.